+++ previous-versions/sync-for-commerce-version-1/accountingcompanyinfo.go
@@ -211,12 +211,12 @@ func (s *accountingCompanyInfo) Refresh(ctx context.Context, request operations.
 	case httpRes.StatusCode == 200:
 		switch {
 		case utils.MatchContentType(contentType, `application/json`):
-			var out *shared.AccountingDataset
+			var out *shared.Dataset
 			if err := utils.UnmarshalJsonFromResponseBody(bytes.NewBuffer(rawBody), &out); err != nil {
 				return nil, err
 			}
 
-			res.AccountingDataset = out
+			res.Dataset = out
 		default:
 			return nil, sdkerrors.NewSDKError(fmt.Sprintf("unknown content-type received: %s", contentType), httpRes.StatusCode, string(rawBody), httpRes)
 		}