+++ previous-versions/sync-for-commerce-version-1/pkg/models/shared/createcompany.go
@@ -3,8 +3,17 @@
 package shared
 
 type CreateCompany struct {
-	// Name of the company in Codat with a partner-commerce data connection.
+	// Additional information about the company. This can be used to store foreign IDs, references, etc.
+	Description *string `json:"description,omitempty"`
+	// Name of company being connected.
 	Name string `json:"name"`
+}
+
+func (o *CreateCompany) GetDescription() *string {
+	if o == nil {
+		return nil
+	}
+	return o.Description
 }
 
 func (o *CreateCompany) GetName() string {