+++ previous-versions/sync-for-commerce-version-1/pushdata.go
@@ -213,12 +213,12 @@ func (s *pushData) ListOperations(ctx context.Context, request operations.GetCom
 	case httpRes.StatusCode == 200:
 		switch {
 		case utils.MatchContentType(contentType, `application/json`):
-			var out *shared.PushHistoryResponse
+			var out *shared.PushOperations
 			if err := utils.UnmarshalJsonFromResponseBody(bytes.NewBuffer(rawBody), &out); err != nil {
 				return nil, err
 			}
 
-			res.PushHistoryResponse = out
+			res.PushOperations = out
 		default:
 			return nil, sdkerrors.NewSDKError(fmt.Sprintf("unknown content-type received: %s", contentType), httpRes.StatusCode, string(rawBody), httpRes)
 		}