+++ previous-versions/sync-for-commerce-version-1/refreshdata.go
@@ -533,12 +533,12 @@ func (s *refreshData) ListPullOperations(ctx context.Context, request operations
 	case httpRes.StatusCode == 200:
 		switch {
 		case utils.MatchContentType(contentType, `application/json`):
-			var out *shared.DataConnectionHistory
+			var out *shared.PullOperations
 			if err := utils.UnmarshalJsonFromResponseBody(bytes.NewBuffer(rawBody), &out); err != nil {
 				return nil, err
 			}
 
-			res.DataConnectionHistory = out
+			res.PullOperations = out
 		default:
 			return nil, sdkerrors.NewSDKError(fmt.Sprintf("unknown content-type received: %s", contentType), httpRes.StatusCode, string(rawBody), httpRes)
 		}