+++ previous-versions/sync-for-commerce-version-1/RELEASES.md
@@ -9,3 +9,13 @@ Based on:
 - [go v0.1.0] previous-versions/sync-for-commerce-version-1
 ### Releases
 - [Go v0.1.0] https://github.com/codatio/client-sdk-go/releases/tag/previous-versions/sync-for-commerce-version-1/v0.1.0 - previous-versions/sync-for-commerce-version-1
+
+## 2023-09-05 00:16:24
+### Changes
+Based on:
+- OpenAPI Doc 1.1 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Sync-Commerce-v1.yaml
+- Speakeasy CLI 1.78.3 (2.96.3) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [go v0.1.1] previous-versions/sync-for-commerce-version-1
+### Releases
+- [Go v0.1.1] https://github.com/codatio/client-sdk-go/releases/tag/previous-versions/sync-for-commerce-version-1/v0.1.1 - previous-versions/sync-for-commerce-version-1